requirements.txt
pygame
numpy
torch
Pillow